+++ components/Tabs/index.scss
@@ -36,7 +36,6 @@ $tabs: 'snake-tabs';
     }
 
     &-item {
-      flex: 1;
       position: relative;
       display: flex;
       justify-content: center;
@@ -44,8 +43,10 @@ $tabs: 'snake-tabs';
       color: #666;
       padding: 6px 12px;
 
-      &-customWidth {
-        flex: initial;
+      /* 禁用态 */
+      &-disabled {
+        color: rgba(0, 0, 0, 0.25);
+        cursor: not-allowed;
       }
     }
 
@@ -87,5 +88,21 @@ $tabs: 'snake-tabs';
   &-left,
   &-right {
     display: flex;
+  }
+
+  &-card {
+    &-top {
+      background: #fafafa;
+      border: 1px solid #e8e8e8;
+      border-radius: 4px 4px 0 0;
+      border-bottom: 1px solid #fff;
+      margin-right: 2px;
+
+      &-current {
+        background: #fff;
+        border-bottom: 1px solid #fff;
+        z-index: 1;
+      }
+    }
   }
 }
