+++ components/index.tsx
@@ -2,7 +2,7 @@ import './styles/index.scss'
 export { default as Breadcrumb } from './Breadcrumb/index'
 export { default as Button } from './Button/index'
 export { default as Panel } from './Panel/index'
-export { default as Tabs } from './Tabs/index'
+export { default as Tabs } from './Tabs/tabs'
 export { default as Icon } from './Icon/index'
 export { default as Checkbox } from './Checkbox/index'
 export { default as Radio } from './Radio/index'
