+++ docs/components/Tabs/README.md
@@ -19,11 +19,11 @@ imports:
 
 <div><Code>{TabsDemoCode}</Code></div>
 
-## Props
+## Api
 
 | 属性 | 说明 | 类型 | 默认值 | 必选 |
 |:-|:-|:-|:-|:-|
-| tabs | tabs 中的展示项 | Array<>, <> 内为对象 title: string | [] | 是 |
+| tabs | tabs 中的展示项 | TabsItemProps[] | [] | 是 |
 | activeTab | tab 当前选中页 | number | 0 | 否 |
 | onChange | Tab 切换时触发 | Function | | 否 |
 | onTabClick | 点击 Tab 时触发 | Function | | 否 |
@@ -31,3 +31,11 @@ imports:
 | tabBarActiveTextColor | tabBar激活Tab文字颜色 | string | | false |
 | tabBarInactiveTextColor | tabBar非激活Tab文字颜色 | string | | false |
 | tabItemStyle | tabItem 的样式 | React.CSSProperties | | false |
+| type | 提供 'line' \| 'card' 两种模式 | string | 'line' | false |
+
+## TabsItemProps Api
+
+| 属性 | 说明 | 类型 | 默认值 | 必选 |
+|:-|:-|:-|:-|:-|
+| title | Tabs 中的单项 | String \| Number \| ReactNode | -- | 是 |
+| disabled | Tabs 中的单项是否禁用 | boolean | false | 否 |
